+++ KAMEditor/gui/commands/arcdialog.h
@@ -17,7 +17,7 @@ class ArcDialog : public QDialog//, private CommandDialog
     Q_OBJECT
 
 public:
-    explicit ArcDialog(CommandsManager* _commandsManager, int _position, QWidget *parent = 0);
+    explicit ArcDialog(CommandsManager* _commandsManager, size_t _position, QWidget *parent = 0);
     ~ArcDialog();
 
 private slots:
@@ -25,7 +25,7 @@ private slots:
 private:
     Ui::ArcDialog *ui;
     CommandsManager* commandsManager;
-    int position;
+    size_t position;
 };
 
 #endif // ARCDIALOG_H
